webapp/requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
scrapy>=2.8.0
scrapy-playwright==0.0.44
playwright>=1.35.0
